requirements.txt
Django~=5.1.6
drf-yasg~=1.21.9
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.5.0
django-cors-headers~=4.7.0
pillow~=11.1.0